Perform Steps on a different thread
I'm changing this because when *stepping* into a simulated function that waits for user input, the wait blocked the invoking SignalR thread, resulting in a deadlock. I would like to come up with a better solution for this particular scenario as I think that consuming a thread from the pool for each step is quite wasteful.